Use VueRouter for proper single page support.
Create a new folder specifically for views (not components). These are the actual rendered pages, while components should contain any shared elements.

Convert all fetch calls to use axion and make these async to avoid callbacks.

Add an interceptor to automatically include authorization if a token is stored, and automatically remove the token when it stops working. This should probably be handled with a bit more care in the future.
