+++ src/main.js
@@ -1,11 +1,10 @@
 import Vue from 'vue'
 import App from '@/App.vue'
 import router from '@/router'
-import API from '@/api'
+import api from '@/api'
 
 Vue.config.productionTip = false
-
-Vue.prototype.$http = API;
+Vue.prototype.$http = api
 
 new Vue({
     router,
